Allow local overwriting of any parameters from config
Enables within script overwriting of any parameters set by the config file by passing a second dictionary with (name, value) as the key value pair. Implemented this in the BaseModel class so it's available for all models. 

Also changed existing code in individual_cv class to be consistent with new interface.