Reusing single model and putting init into Base
I put the model construction into the BaseModel class since we are now using that single model object in the same way for all our models.

For the IndividualModel class, now instead of saving models, we just save the parameters and load them into self.model whenever needed. This prevents the memory issue of having multiple graphs and gives a substantial speedup.