Refactoring evaluate method
Moved the core of the evaluate method to the BaseModel class since it is a method we want to be common to all subclasses. Adjusted predict methods so they are consistent as well, and added get_score methods since they are also unique to each subclass. Also adjusted experiments for the changes made